Added ability to pass command-line options to ffmpeg.

+++ Scripts/encodeDir.sh
@@ -3,11 +3,11 @@
 # encodeDir.sh
 # ndt: n-dimensional tracer
 #
-# Copyright (c) 2014-2020 Bryan Franklin. All rights reserved.
+# Copyright (c) 2014-2021 Bryan Franklin. All rights reserved.
 
 if [ -z "$1" ]; then
     echo "Usage:"
-    echo "  $0 directory [fps]"
+    echo "  $0 directory [fps] [options]"
     exit 1
 fi
 
@@ -28,6 +28,12 @@ else
     VIDEO=videos/${NAME}_${FPS}fps.avi
 fi
 
+# remove processed command-line options
+shift 2
+if [ ! -z "$*" ]; then
+    echo "ffmpeg options: $*"
+fi
+
 echo $VIDEO
 
 mkdir -p videos
@@ -43,7 +49,8 @@ mencoder "mf://$DIR/*.png" -mf fps=$FPS -oac copy -ovc x264 -x264encopts pass=1:
 mencoder "mf://$DIR/*.png" -mf fps=$FPS -oac copy -ovc x264 -x264encopts pass=2:preset=veryslow:fast_pskip=0:tune=film:frameref=15:bitrate=3000:threads=auto -o $VIDEO
 
 MP4=`echo $VIDEO | sed 's/.avi$/.mp4/'`
-rm -f "$MP4"
-ffmpeg -i "$VIDEO" "$MP4"
+test -f "$MP4" && rm -f "$MP4"
+echo "ffmpeg -i $VIDEO $* $MP4"
+ffmpeg -i "$VIDEO" $* "$MP4"
 
-echo $MP4
+ls -l $MP4
